chore: move cmds to cmd package
